rename autocalculated fields and trigger validations
