+++ sdk/src/main/scala/com/stratio/sparta/sdk/AutoCalculatedField.scala
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package com.stratio.sparta.serving.core.models
+package com.stratio.sparta.sdk
 
-case class AutoCalculateFieldModel(
-                                    fromNotNullFields: Option[FieldModel] = None,
-                                    fromPkFields: Option[FieldModel] = None,
-                                    fromFields: Option[FromFieldModel] = None,
-                                    fromFixedValue: Option[FromFixedValueModel] = None
-                                  )
+case class AutoCalculatedField(
+                               fromNotNullFields: Option[Field],
+                               fromPkFields: Option[Field],
+                               fromFields: Option[FromField],
+                               fromFixedValue: Option[FromFixedValue]
+                             )
